# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music app powered by yt-dlp
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Track {
  id: string;
  title: string;
  /** @nullable */
  uploader?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  viewCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
  /** @nullable */
  uploadDate?: string | null;
  webpage_url: string;
}

export interface TrackDetail {
  id: string;
  title: string;
  /** @nullable */
  uploader?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  viewCount?: number | null;
  /** @nullable */
  likeCount?: number | null;
  /** @nullable */
  commentCount?: number | null;
  /** @nullable */
  uploadDate?: string | null;
  tags: string[];
  categories: string[];
  webpage_url: string;
  /** @nullable */
  channelId?: string | null;
  /** @nullable */
  channelUrl?: string | null;
}

export interface HomeSection {
  id: string;
  title: string;
  subtitle: string;
  query: string;
  tracks: Track[];
}

export interface HomeFeed {
  region: string;
  sections: HomeSection[];
}

export interface StreamUrlResult {
  videoId: string;
  url: string;
  ext: string;
  /** @nullable */
  acodec?: string | null;
  /** @nullable */
  abr?: number | null;
  /** @nullable */
  expiry?: string | null;
}

export interface AudioFormat {
  formatId: string;
  ext: string;
  /** @nullable */
  acodec?: string | null;
  /** @nullable */
  abr?: number | null;
  /** @nullable */
  asr?: number | null;
  /** @nullable */
  filesize?: number | null;
  url: string;
  /** @nullable */
  note?: string | null;
}

export interface Playlist {
  id: string;
  title: string;
  /** @nullable */
  uploader?: string | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  description?: string | null;
  trackCount: number;
  tracks: Track[];
}

export interface SubtitleLanguage {
  language: string;
  name: string;
  ext: string;
}

export interface CookiesStatus {
  hasCookies: boolean;
  message?: string;
}

export interface LrcLine {
  time: number;
  text: string;
}

export interface LyricsResult {
  found: boolean;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  trackName?: string | null;
  /** @nullable */
  artistName?: string | null;
  /** @nullable */
  plainLyrics?: string | null;
  syncedLyrics?: LrcLine[] | null;
}

export interface SubtitlesResult {
  videoId: string;
  subtitles: SubtitleLanguage[];
  automatic: SubtitleLanguage[];
}

export type SearchTracksParams = {
/**
 * Search query
 */
q: string;
/**
 * Max number of results
 */
limit?: number;
};

export type GetTrendingParams = {
limit?: number;
};

export type GetLyricsParams = {
title?: string;
artist?: string;
};